+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 ### Added
 - Added offline for product batch list and detail
 - Added a Capacitations view for Disabilities List
+- Added offline for supply batches
 
 ## [0.2.0] - 29/11/2025